// ui-parts/carousel/src/pagination/appearance/appearance.interfaces.ts
export interface PaginationAppearanceStyles {
  backgroundColor?: string | Function
  fontColor?: string | Function
  borderColor?: string | Function
  activeFontColor?: string | Function
  activeBackgroundColor?: string | Function
  activeBorderColor?: string | Function
  borderStyle?: string | Function
}